refactor : split remove from useData logic
