+++ src/pages/search/index.tsx
@@ -1,24 +1,13 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import Head from 'next/head';
 import { PostList } from 'components/Post';
-import {
-  useUser,
-  useAllReviews,
-  useIntersectionObserver,
-  useSinglePostModal,
-} from 'hooks';
-import { REQUEST, SUCCESS } from 'hooks/common/useApiFetch';
+import { useUser, useAllReviews, useIntersectionObserver } from 'hooks';
+import { REQUEST } from 'hooks/common/useApiFetch';
 import { Loading } from 'components/UI';
 import styled from '@emotion/styled';
 import dynamic from 'next/dynamic';
 
-const Modal = dynamic(() => import('components/UI/Modal'));
-const Card = dynamic(() => import('components/UI/Card'));
-const LoadingSinglePost = dynamic(
-  () => import('components/UI/LoadingSinlgePost')
-);
 const WriteButton = dynamic(() => import('components/Post/WriteButton'));
-const SinglePost = dynamic(() => import('components/Post/SinglePost'));
 
 const SearchMain = () => {
   const { user } = useUser();
@@ -26,7 +15,7 @@ const SearchMain = () => {
   const [
     allReviews,
     fetchReviewHanlder,
-    fetchRemoveHanlder,
+    removeCacheHandler,
     allReviewsFetchStatus,
     hasMore,
     lastKey,
@@ -36,17 +25,6 @@ const SearchMain = () => {
     deps: [hasMore, lastKey],
     fetcher: fetchReviewHanlder,
   });
-
-  const modalController = useSinglePostModal(allReviews);
-
-  const removeHanlder = useCallback(
-    (id: string) => () => {
-      modalController.closeModal(); // close Modal
-      modalController.removeCache(id); // remove Cache
-      fetchRemoveHanlder(id);
-    },
-    [modalController]
-  );
 
   return (
     <>
@@ -65,31 +43,9 @@ const SearchMain = () => {
       </Head>
       <PostList
         reviewData={allReviews}
-        openSinglePost={modalController.openModal}
+        removeCacheFromDataHandler={removeCacheHandler}
       />
       {user && user.isLoggedIn && <WriteButton />}
-      {modalController.showModal && (
-        <Modal modalHandler={modalController.closeModal}>
-          <Card isModal={true}>
-            {!modalController.singleReview ||
-            modalController.singleReviewFetchStatus === REQUEST ||
-            modalController.singleReviewFetchStatus === SUCCESS ? (
-              <LoadingSinglePost />
-            ) : (
-              <SinglePost
-                data={modalController.singleReview}
-                NavigationInfo={{
-                  hasPrev: modalController.index > 0,
-                  hasNext: modalController.index < allReviews.length - 1,
-                  prevHandler: modalController.prevHandler,
-                  nextHandler: modalController.nextHandler,
-                }}
-                removeHanlder={removeHanlder}
-              />
-            )}
-          </Card>
-        </Modal>
-      )}
       <LoaderContainer ref={observerTarget}>
         {allReviewsFetchStatus === REQUEST && <Loading />}
       </LoaderContainer>
